% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/argmin.R
\name{argmin}
\alias{argmin}
\title{argmin Argmin function for a vector.}
\usage{
argmin(v, target = 0)
}
\arguments{
\item{v}{The vector to compare to target.}

\item{target}{The value sought in the vector; default=0.}
}
\value{
The index in v of the value which is closest to target.
}
\description{
Return the index minimizing distance from v to target.
}

