% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binom.confint.R
\name{binom.confint}
\alias{binom.confint}
\title{binom.confint}
\usage{
binom.confint(k, n, alpha = 0.05, side = c("two", "upper", "lower"))
}
\arguments{
\item{k}{#"heads"}

\item{n}{sample size}

\item{alpha}{Confidence level}

\item{side}{Sidedness of the hypothesis: c("two", "upper", "lower")}
}
\description{
Exact confidence intervals for a binomial proportion parameter.
}

