% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sesp-pv-NNT.R
\name{NNT.from.pv}
\alias{NNT.from.pv}
\alias{pv.to.NNT}
\title{NNT.from.pv}
\usage{
NNT.from.pv(ppv, npv, pv)
}
\arguments{
\item{ppv}{Positive predictive value}

\item{npv}{Negative predictive value}

\item{pv}{Alternative input of ppv and npv, in matrix or named vector.}
}
\value{
c(NNTpos=NNTpos, NNTneg=NNTneg)
}
\description{
Compute NNT values from predictive values.
}

