% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sesp-pv-NNT.R
\name{NNT.from.sesp}
\alias{NNT.from.sesp}
\alias{sesp.to.NNT}
\title{NNT.from.sesp}
\usage{
NNT.from.sesp(se, sp, sesp, prev)
}
\arguments{
\item{se}{Positive predictive value}

\item{sp}{Negative predictive value}

\item{sesp}{Alternative input for se and sp, as matrix or named vector.}

\item{prev}{Prevalence (prior probability)}
}
\value{
c(NNTpos=NNTpos, NNTneg=NNTneg)
}
\description{
Compute NNT values from sensitivity aand specificity.
}

