% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sesp-pv-NNT.R
\name{NNT.to.pv}
\alias{NNT.to.pv}
\alias{pv.from.NNT}
\title{NNT.to.pv}
\usage{
NNT.to.pv(NNTpos, NNTneg, NNT, prev, calculate.se.sp = F)
}
\arguments{
\item{NNTpos}{NNT for a patient with positive test result}

\item{NNTneg}{NNT for a patient with negative test result}

\item{NNT}{A matrix or vector of (NNTpos, NNTneg) values.}

\item{prev}{Prevalence of the "BestToTreat" group before testing.}

\item{calculate.se.sp}{(default=FALSE) If TRUE, also calculate the sensitivity and specificity using the contra-Bayes theorem.}
}
\value{
For matrix input, cbind(ppv=ppv, npv=npv). For vector input, c(ppv=ppv, npv=npv).
}
\description{
Convert between (NNTpos, NNTneg)  and (PPV, NPV).
}

