% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sesp-pv-NNT.R
\name{NNT.to.sesp}
\alias{NNT.to.sesp}
\alias{sesp.from.NNT}
\title{NNT.to.sesp}
\usage{
NNT.to.sesp(NNTpos, NNTneg, NNT, prev)
}
\arguments{
\item{NNTpos}{NNT for a positive test result}

\item{NNTneg}{NNT for a negative test result}

\item{NNT}{Alternative way in input NNT values (matrix or vector)}

\item{prev}{Prevalence (prior probability)}
}
\value{
c(se=se, sp=sp)
}
\description{
Compute sensitivity aand specificity from NNT values.
}

