% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sesp-pv-NNT.R
\name{pv.from.sesp}
\alias{pv.from.sesp}
\alias{sesp.to.pv}
\title{pv.from.sesp}
\usage{
pv.from.sesp(se = 0.8, sp = 0.8, sesp, prev = 0.001)
}
\arguments{
\item{se}{Positive predictive value}

\item{sp}{Negative predictive value}

\item{sesp}{Alternative input for se and sp, as matrix or named vector.}

\item{prev}{Prevalence (prior probability). Default =  0.001}
}
\value{
c(ppv=ppv, npv=npv)
}
\description{
Computes predictive values from sensitivity and specificity.
}

