% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{run}
\usage{
run(shinyDir)
}
\arguments{
\item{shinyDir}{Current options are "shinyElicit" and "shinyCombinePlots". If not provided, a menu of the options is provided.}
}
\description{
Run a shiny app for this package.
}
\details{
The selected shiny app is run. See the vignette Using_the_NNTbiomarker_package for details, and the vignette The_Biomarker_Crisis for an overview.
}

