% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run.R
\name{runCombinePlots}
\alias{runCombinePlots}
\title{runCombinePlots}
\usage{
runCombinePlots()
}
\description{
Run a shiny app connecting a visual scale for NNT quantities and a "contra-Bayes" plot for mapping from predictive values to sensitivity/specificity (Bayes theorem in reverse).
}
\seealso{
run
}

