% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sesp-pv-NNT.R
\name{sesp.from.pv}
\alias{pv.to.sesp}
\alias{sesp.from.pv}
\title{sesp.from.pv}
\usage{
sesp.from.pv(ppv = 0.1, npv = 0.7, pv, prev = 0.2)
}
\arguments{
\item{ppv}{Positive predictive value}

\item{npv}{Negative predictive value}

\item{pv}{Alternative input of ppv and npv, in matrix or named vector.}

\item{prev}{Prevalence (prior probability)}
}
\value{
c(se=se,sp=sp)
}
\description{
Computes sensitivity and specificity from  predictive values.
}

