% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ifVerboseCat.R
\name{setVerboseCatOption}
\alias{ifVerboseCat}
\alias{setVerboseCatOption}
\title{setVerboseCatOption}
\usage{
setVerboseCatOption(fname, value)
}
\arguments{
\item{fname}{Name of the function to control.}

\item{value}{Boolean value: should this function print out messages?}
}
\value{
The new value of the namespace option for fname ifVerboseCat
}
\description{
Allows user to toggle on and off printing messages on a per-function basis.
Should be usable in other packages, but not by importing.
}

